KVM: Add PV framework
To get proper performance out of PR=1 virtual machines on PPC we need to have
a way to communicate with the hypervisor without trapping. MOL does this using
a shared page for pr_regs. KVM does this using a shared page for safe guest
registers.

This patch adds code to make use of the PV framwork, speeding up the guest
by a _lot_. Keep in mind that this will only be part of 2.6.37, as the
framework is really new currently.

Signed-off-by: Alexander Graf <agraf@suse.de>